third_party/blink/web_tests/dom/document_type/next-element-sibling.html

<!DOCTYPE html>
<title>This tests the NonDocumentTypeChildNode interface nextElementSibling property on a DocumentType node</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
setup({single_test: true});
var doctype = document.doctype;
assert_equals(doctype.nextElementSibling, undefined);
done();
</script>
